♻️  Refact: LocationService 프로토콜을 Common/Protocol 폴더로 분리 & CLAuthorizationStatus 확장 기능 추가
