+++ EatPic-iOS/Sources/Core/Location/LocationServiceImpl.swift
@@ -1,5 +1,5 @@
 //
-//  LocationService.swift
+//  LocationServiceImpl.swift
 //  EatPic-iOS
 //
 //  Created by jaewon Lee on 7/15/25.
@@ -7,17 +7,6 @@
 
 import Foundation
 import CoreLocation
-import SwiftUI
-
-/// 위치 정보를 외부에 제공하는 추상화 인터페이스
-protocol LocationService {
-    var currentLocation: CLLocation? { get }
-    var authorizationStatus: CLAuthorizationStatus { get }
-    
-    func requestAuthorization()
-    func startUpdatingLocation()
-    func stopUpdatingLocation()
-}
 
 final class LocationServiceImpl: NSObject, LocationService {
     private let manager = CLLocationManager()
