+++ EatPic-iOS/Sources/Core/Location/LocationViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  LocationViewModel.swift
+//  EatPic-iOS
+//
+//  Created by jaewon Lee on 7/15/25.
+//
+
+import Foundation
